requirements.txt
user-agents
pydantic
fastapi
python-dotenv
pydantic-settings
sqlalchemy
